refactor(cli): flatten host command module
Move the Host command slice to the neutral command root and update its internal export and test-only temporary prefixes without changing public CLI or host API behavior.
